Categorize failing ApApFunFun tests
The problem is patterns like ApApFunFun (or ApBindFunFun).  Each Fun
level removes any ExVars established lower than it from the context,
which means that the second Fun level can never be solved if it isn't
solved at the bottom.

This is not a problem in the bidir paper rules because the return type
of a Fun level gets its own ExVar which substitutes in for the lower
level ExVars that are erased (more specifically, new ExVars are
inserted *before* its own ExVar to cover them).

It seems that the most straightforward way to address this is to add
computation ExVars to the context, so that the bidir paper rules can
be implemented more closely.