Add (failing) CompSub tests
Computations that introduce existential index variables incorrectly
check against any type, because the existentials are allowed to solve
to different values in different subcheck contexts.

In the DML rules, these subchecks would be collected into a single
large constraint that must be solved all together.

The solution is to not actually existentially quantify index
variables, but rather solve-and-replace in the same way that
existential type variables are solved.